requirements.txt
ncatbot
openai
aiosqlite
markdown-it-py
playwright
